reorganize some components/routes
